+++ CHANGELOG.rst
@@ -16,6 +16,17 @@ Unreleased
 
 *
 
+[3.0.0]
+~~~~~~~
+
+* **Breaking change**: Caliper and xAPI processors will reject events unless the ``CALIPER_EVENTS_ENABLED`` and ``XAPI_EVENTS_ENABLED`` Django settings are enabled, respectively.
+
+
+[2.0.0]
+~~~~~~~
+
+(no changelog recorded)
+
 [1.0.0] - 2020-11-09
 ~~~~~~~~~~~~~~~~~~~~